requirements.txt
psutil
python-dotenv
smtplib
